Move policy replicas from DDMClient to PolicySource
